+++ Keter/Proxy.hs
@@ -5,10 +5,14 @@
 -- | A light-weight, minimalistic reverse HTTP proxy.
 module Keter.Proxy
     ( reverseProxy
-    , HostLookup
+    , makeSettings
+    , ProxySettings(..)
     , TLSConfig (..)
     ) where
 
+import qualified Network.HTTP.Conduit      as HTTP
+import qualified Data.CaseInsensitive      as CI
+import qualified Keter.HostManager         as HostMan
 import           Blaze.ByteString.Builder          (copyByteString)
 import           Control.Applicative               ((<$>), (<|>))
 import           Control.Monad.IO.Class            (liftIO)
@@ -59,6 +63,7 @@ import           Network.Wai.Middleware.Gzip       (gzip, GzipSettings(..), Gzip
 import           Prelude                           hiding (FilePath, (++))
 import           WaiAppStatic.Listing              (defaultListing)
 import qualified Network.TLS as TLS
+import qualified System.Directory as Dir
 
 #if !MIN_VERSION_http_reverse_proxy(0,6,0)
 defaultWaiProxySettings = def
@@ -69,26 +74,51 @@ defaultLocalWaiProxySettings = def
 #endif
 
 
--- | Mapping from virtual hostname to port number.
-type HostLookup = ByteString -> IO (Maybe (ProxyAction, TLS.Credentials))
+data ProxySettings = MkProxySettings
+  -- | Mapping from virtual hostname to port number.
+  { psHostLookup  :: ByteString -> IO (Maybe (ProxyAction, TLS.Credentials))
+  , psManager     :: !Manager
+  , psConfig      :: !KeterConfig
+  , psUnkownHost  :: ByteString -> ByteString
+  , psMissingHost :: ByteString
+  }
 
-reverseProxy :: Bool
-             -> Int -> Manager -> HostLookup -> ListeningPort -> IO ()
-reverseProxy useHeader timeBound manager hostLookup listener =
-    run $ gzip def{gzipFiles = GzipPreCompressed GzipIgnore} $ withClient isSecure useHeader timeBound manager hostLookup
+makeSettings :: KeterConfig -> HostMan.HostManager -> IO ProxySettings
+makeSettings psConfig@KeterConfig {..} hostman = do
+    psManager <- HTTP.newManager HTTP.tlsManagerSettings
+    psMissingHost <- case kconfigMissingHostResponse of
+      Nothing -> pure defaultMissingHostBody
+      Just x -> taggedReadFile "unknown-host-response-file" x
+    psUnkownHost <- case kconfigUnknownHostResponse  of
+                Nothing -> pure defaultUnknownHostBody
+                Just x -> const <$> taggedReadFile "missing-host-response-file" x
+    pure $ MkProxySettings{..}
+    where
+        psHostLookup = HostMan.lookupAction hostman . CI.mk
+
+taggedReadFile :: String -> FilePath -> IO ByteString
+taggedReadFile tag file = do
+        isExist <- Dir.doesFileExist file
+        if isExist then S.readFile file else do
+          wd <- Dir.getCurrentDirectory
+          error $ "could not find " <> tag <> " on path '" <> file <> "' with working dir '" <> wd <> "'"
+
+reverseProxy :: ProxySettings -> ListeningPort -> IO ()
+reverseProxy settings listener =
+    run $ gzip def{gzipFiles = GzipPreCompressed GzipIgnore} $ withClient isSecure settings
   where
     warp host port = Warp.setHost host $ Warp.setPort port Warp.defaultSettings
     (run, isSecure) =
         case listener of
             LPInsecure host port -> (Warp.runSettings (warp host port), False)
             LPSecure host port cert chainCerts key session -> (WarpTLS.runTLS
-                (connectClientCertificates hostLookup session $ WarpTLS.tlsSettingsChain
+                (connectClientCertificates (psHostLookup settings) session $ WarpTLS.tlsSettingsChain
                     cert
                     (V.toList chainCerts)
                     key)
                 (warp host port), True)
 
-connectClientCertificates :: HostLookup -> Bool -> WarpTLS.TLSSettings -> WarpTLS.TLSSettings
+connectClientCertificates :: (ByteString -> IO (Maybe (ProxyAction, TLS.Credentials))) -> Bool -> WarpTLS.TLSSettings -> WarpTLS.TLSSettings
 connectClientCertificates hl session s =
     let
         newHooks@TLS.ServerHooks{..} = WarpTLS.tlsServerHooks s
@@ -102,12 +132,9 @@ connectClientCertificates hl session s =
           , WarpTLS.tlsSessionManagerConfig = if session then (Just TLSSession.defaultConfig) else Nothing }
 
 withClient :: Bool -- ^ is secure?
-           -> Bool -- ^ use incoming request header for IP address
-           -> Int  -- ^ time bound for connections
-           -> Manager
-           -> HostLookup
+           -> ProxySettings
            -> Wai.Application
-withClient isSecure useHeader bound manager hostLookup =
+withClient isSecure MkProxySettings {..} =
     waiProxyToSettings
        (error "First argument to waiProxyToSettings forced, even thought wpsGetDest provided")
        defaultWaiProxySettings
@@ -116,8 +143,15 @@ withClient isSecure useHeader bound manager hostLookup =
                 then SIHFromHeader
                 else SIHFromSocket
         ,  wpsGetDest = Just getDest
-        } manager
+        } psManager
   where
+    useHeader :: Bool
+    useHeader = kconfigIpFromHeader psConfig
+
+    -- calculate the number of microseconds since the
+    -- configuration option is in milliseconds
+    bound :: Int
+    bound = kconfigConnectionTimeBound psConfig * 1000
     protocol
         | isSecure = "https"
         | otherwise = "http"
@@ -140,7 +174,8 @@ withClient isSecure useHeader bound manager hostLookup =
     getDest :: Wai.Request -> IO (LocalWaiProxySettings, WaiProxyResponse)
     getDest req =
         case Wai.requestHeaderHost req of
-            Nothing -> return (defaultLocalWaiProxySettings, WPRResponse missingHostResponse)
+            Nothing -> do
+              return (defaultLocalWaiProxySettings, WPRResponse $ missingHostResponse psMissingHost)
             Just host -> processHost req host
 
     processHost :: Wai.Request -> S.ByteString -> IO (LocalWaiProxySettings, WaiProxyResponse)
@@ -148,16 +183,17 @@ withClient isSecure useHeader bound manager hostLookup =
         -- Perform two levels of lookup. First: look up the entire host. If
         -- that fails, try stripping off any port number and try again.
         mport <- liftIO $ do
-            mport1 <- hostLookup host
+            mport1 <- psHostLookup host
             case mport1 of
                 Just _ -> return mport1
                 Nothing -> do
                     let host' = S.takeWhile (/= 58) host
                     if host' == host
                         then return Nothing
-                        else hostLookup host'
+                        else psHostLookup host'
         case mport of
-            Nothing -> return (defaultLocalWaiProxySettings, WPRResponse $ unknownHostResponse host)
+            Nothing -> do
+              return (defaultLocalWaiProxySettings, WPRResponse $ unknownHostResponse host (psUnkownHost host))
             Just ((action, requiresSecure), _)
                 | requiresSecure && not isSecure -> performHttpsRedirect host req
                 | otherwise -> performAction req action
@@ -190,7 +226,10 @@ withClient isSecure useHeader bound manager hostLookup =
             })
     performAction req (PARedirect config) = return (addjustGlobalBound Nothing, WPRResponse $ redirectApp config req)
     performAction _ (PAReverseProxy config rpconfigMiddleware tbound) =
-       return (addjustGlobalBound tbound, WPRApplication $ processMiddleware rpconfigMiddleware $ Rewrite.simpleReverseProxy manager config)
+       return (addjustGlobalBound tbound, WPRApplication
+                $ processMiddleware rpconfigMiddleware
+                $ Rewrite.simpleReverseProxy psManager config
+              )
 
 redirectApp :: RedirectConfig -> Wai.Request -> Wai.Response
 redirectApp RedirectConfig {..} req =
@@ -233,16 +272,22 @@ redirectApp RedirectConfig {..} req =
         , Wai.rawQueryString req
         ]
 
-missingHostResponse :: Wai.Response
-missingHostResponse = Wai.responseBuilder
-    status200
-    [("Content-Type", "text/html; charset=utf-8")]
-    $ copyByteString "<!DOCTYPE html>\n<html><head><title>Welcome to Keter</title></head><body><h1>Welcome to Keter</h1><p>You did not provide a virtual hostname for this request.</p></body></html>"
+defaultMissingHostBody :: ByteString
+defaultMissingHostBody = "<!DOCTYPE html>\n<html><head><title>Welcome to Keter</title></head><body><h1>Welcome to Keter</h1><p>You did not provide a virtual hostname for this request.</p></body></html>"
 
-unknownHostResponse :: ByteString -> Wai.Response
-unknownHostResponse host = Wai.responseBuilder
+missingHostResponse :: ByteString -> Wai.Response
+missingHostResponse missingHost = Wai.responseBuilder
     status200
     [("Content-Type", "text/html; charset=utf-8")]
-    (copyByteString "<!DOCTYPE html>\n<html><head><title>Welcome to Keter</title></head><body><h1>Welcome to Keter</h1><p>The hostname you have provided, <code>"
-     `mappend` copyByteString host
-     `mappend` copyByteString "</code>, is not recognized.</p></body></html>")
+    $ copyByteString missingHost
+
+defaultUnknownHostBody :: ByteString -> ByteString
+defaultUnknownHostBody host =
+  "<!DOCTYPE html>\n<html><head><title>Welcome to Keter</title></head><body><h1>Welcome to Keter</h1><p>The hostname you have provided, <code>"
+  <> host <> "</code>, is not recognized.</p></body></html>"
+
+unknownHostResponse :: ByteString -> ByteString -> Wai.Response
+unknownHostResponse host body = Wai.responseBuilder
+    status200
+    [("Content-Type", "text/html; charset=utf-8"), ("X-Forwarded-Host", host)]
+    (copyByteString body)