+++ Keter/Types/V10.hs
@@ -105,6 +105,9 @@ data KeterConfig = KeterConfig
     -- ^ Maximum request time in milliseconds per connection.
     , kconfigCliPort             :: !(Maybe Port)
     -- ^ Port for the cli to listen on
+
+    , kconfigUnknownHostResponse  :: !(Maybe F.FilePath)
+    , kconfigMissingHostResponse :: !(Maybe F.FilePath)
     }
 
 instance ToCurrent KeterConfig where
@@ -121,6 +124,8 @@ instance ToCurrent KeterConfig where
         , kconfigEnvironment = Map.empty
         , kconfigConnectionTimeBound = connectionTimeBound
         , kconfigCliPort             = Nothing
+        , kconfigUnknownHostResponse  = Nothing
+        , kconfigMissingHostResponse = Nothing
         }
       where
         getSSL Nothing = V.empty
@@ -145,6 +150,8 @@ instance Default KeterConfig where
         , kconfigEnvironment = Map.empty
         , kconfigConnectionTimeBound = V04.fiveMinutes
         , kconfigCliPort             = Nothing
+        , kconfigUnknownHostResponse  = Nothing
+        , kconfigMissingHostResponse = Nothing
         }
 
 instance ParseYamlFile KeterConfig where
@@ -166,6 +173,8 @@ instance ParseYamlFile KeterConfig where
             <*> o .:? "env" .!= Map.empty
             <*> o .:? "connection-time-bound" .!= V04.fiveMinutes
             <*> o .:? "cli-port"
+            <*> o .:? "missing-host-response-file"
+            <*> o .:? "unknown-host-response-file"
 
 -- | Whether we should force redirect to HTTPS routes.
 type RequiresSecure = Bool